refac: split PU gen script into its module
